Split Android cloud HTTP transport helpers
